Update utils.py
This likely happens in the utils.minimize_mask function within the Mask R-CNN library when it tries to resize the mask using skimage.transform.resize. The mask is expected to be a numerical array representing pixel values, but it is currently a boolean array. This means the mask likely contains only True/False values, representing whether each pixel is part of the mask or not.